CC_GWAS

This is a function to run a case-case GWAS using the framework introduced in the ReACt package. The method will first filter out any SNPs with NA in the case or control AF. It will then remove any SNPs that are not present in both datasets. Finally, it will remove duplicate SNPs, by defualt keeping the first record. Then the CC-GWAS will be run.

CC_GWAS(
  trait1_data,
  trait2_data,
  AF_case_colname = "AF_case",
  AF_control_colname = "AF_control",
  trait1_N_case,
  trait1_N_control,
  trait2_N_case,
  trait2_N_control,
  chromosome_colname = "chr",
  position_colname = "pos",
  id_colname = NA,
  A1_colname = "A1",
  A2_colname = "A2",
  rsid_colname = "RSID",
  OR_colname = "OR",
  SE_colname = "SE",
  N_caseInCase = NA,
  N_controlInControl = NA,
  N_caseInControl = NA,
  Zthreshold = 0
)

Arguments

trait1_data

a dataframe containing the chromosome, position, case AF, and control AF for each SNP for trait 1

trait2_data

a dataframe containing the chromosome, position, case AF, and control AF for each SNP for trait 2

AF_case_colname

a string with the name for the column containing the case AF in both trait 1 and trait 2 datasets; default: "AF_case"

trait1_N_case

the number of cases in trait 1 dataset

trait1_N_control

the number of controls in trait 1 dataset

trait2_N_case

the number of cases in trait 2 dataset

trait2_N_control

the number of controls in trait 2 dataset

chromosome_colname

a string with the name for the column containing the SNP chromosome number in both trait 1 and trait 2 datasets; default: "chr"

position_colname

a string with the name for the column containing the SNP position in both trait 1 and trait 2 datasets; default: "pos"

id_colname

a string containing the id for each SNP in both trait 1 and trait 2 datasets; default: NA. If NA, then an ID column will be added and used for data filtering

OR_colname

a string with the name for the column with the OR from the GWAS summary statistics for each trait

SE_colname

a string with the name for the column with the SE from the GWAS summary statistics for each trait

N_caseInCase

a 2x2 matrix where the upper left cell (a) is the number of cases in study 1 and the lower right cell (d) is the number of cases in study 2. Cells b and c are the number of overlapping cases; Default = NA - if left as NA algorithm will assume no sample overlap

N_controlInControl

a 2x2 matrix where the upper left cell (a) is the number of controls in study 1 and the lower right cell (d) is the number of controls in study 2. Cells b and c are the number of overlapping controls; Default = NA - if left as NA algorithm will assume no sample overlap

N_caseInControl

a 2x2 matrix of the number of cases in the control sample. The upper right cell (b) represents the number of cases from trait 1 that are present as controls for trait 2. The lower left cell (c) is the number of cases from trait 2 that are included as controls for trait 1. Cells a and d should be 0; Default = NA - if left as NA algorithm will assume no sample overlap

Zthreshold

value for the z-statistic threshold. If Zthreshold = 0 then will not perform correction for sample overlap (In order to run correction for sample overlap, OR and SE in each trait's data is required). Yang et al suggest using a threshold of z=1.0; however, this will overlap user input sample sizes. If exact sample sizes and overlap is known, this parameter should be left at 0.

AF_control_colnamea

a string with the name for the column containing the control AF in both trait 1 and trait 2 datasets; default: "AF_control"

Value

returns a dataframe with the original data (where each dataset's original columns with have .trait1 or .trait2 appended to its column name) and three additional columns for the CC-GWAS results: OR, SE, p equal to the number of variants

Note

trait1_data and trait2_data need to have the same column names for chromosome, position, OR, SE, and case and control AFs.